+++ raiden/src/store/index.ts
@@ -1,3 +1,2 @@
 export * from './state';
 export { raidenReducer } from './reducers';
-export { raidenEpics } from './epics';
