+++ raiden/tests/e2e/raiden.spec.ts
@@ -14,10 +14,10 @@ import { request } from 'matrix-js-sdk';
 
 import { Raiden } from 'raiden/raiden';
 import { ShutdownReason } from 'raiden/constants';
-import { initialState } from 'raiden/store';
+import { initialState } from 'raiden/store/state';
 import { raidenShutdown } from 'raiden/store/actions';
 import { newBlock } from 'raiden/channels/actions';
-import { ChannelState } from 'raiden/channels';
+import { ChannelState } from 'raiden/channels/types';
 import { Storage } from 'raiden/utils/types';
 import { ContractsInfo, RaidenContracts } from 'raiden/types';
